+++ src/main/java/io/github/astrapi69/json/JsonFileToObjectExtensions.java
@@ -26,6 +26,9 @@ package io.github.astrapi69.json;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -41,6 +44,55 @@ public final class JsonFileToObjectExtensions
 {
 	private JsonFileToObjectExtensions()
 	{
+	}
+
+	/**
+	 * Transforms the given json file into a java map object
+	 *
+	 * @param <K>
+	 *            the generic type of keys
+	 * @param <V>
+	 *            the generic type of values
+	 * @param jsonFile
+	 *            the json file
+	 * @param typeReference
+	 *            the type reference
+	 * @return the t
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred
+	 */
+	public static <K, V> Map<K, V> toMapObject(final File jsonFile,
+		final TypeReference<Map<K, V>> typeReference) throws IOException
+	{
+		Objects.requireNonNull(jsonFile);
+		Objects.requireNonNull(typeReference);
+		return toMapObject(jsonFile, typeReference, ObjectMapperFactory.newObjectMapper(true));
+	}
+
+	/**
+	 * Transforms the given json file into a java map object
+	 *
+	 * @param <K>
+	 *            the generic type of keys
+	 * @param <V>
+	 *            the generic type of values
+	 * @param jsonFile
+	 *            the json file
+	 * @param typeReference
+	 *            the type reference
+	 * @param mapper
+	 *            the object mapper
+	 * @return the t
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred
+	 */
+	public static <K, V> Map<K, V> toMapObject(final File jsonFile,
+		final TypeReference<Map<K, V>> typeReference, final ObjectMapper mapper) throws IOException
+	{
+		Objects.requireNonNull(jsonFile);
+		Objects.requireNonNull(typeReference);
+		Objects.requireNonNull(mapper);
+		return mapper.readValue(jsonFile, typeReference);
 	}
 
 	/**
@@ -107,6 +159,54 @@ public final class JsonFileToObjectExtensions
 		Objects.requireNonNull(typeReference);
 		Objects.requireNonNull(mapper);
 		return mapper.readValue(jsonFile, typeReference);
+	}
+
+	/**
+	 * Transforms the given json file into a java object {@link Collection}
+	 *
+	 * @param <T>
+	 *            the generic type of the return type
+	 * @param jsonFile
+	 *            the json file
+	 * @param collectionClass
+	 *            the collection class
+	 * @param elementClass
+	 *            the element class
+	 * @return the list with the java objects.
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred
+	 */
+	public static <T> Collection<T> toObjectCollection(final File jsonFile,
+		@SuppressWarnings("rawtypes") final Class<? extends Collection> collectionClass,
+		final Class<T> elementClass) throws IOException
+	{
+		Objects.requireNonNull(jsonFile);
+		Objects.requireNonNull(collectionClass);
+		Objects.requireNonNull(elementClass);
+		final ObjectMapper mapper = ObjectMapperFactory.newObjectMapper(true);
+		return mapper.readValue(jsonFile,
+			mapper.getTypeFactory().constructCollectionType(collectionClass, elementClass));
+	}
+
+	/**
+	 * Transforms the given json string into a java object {@link List}
+	 *
+	 * @param <T>
+	 *            the generic type of the return type
+	 * @param jsonFile
+	 *            the json file
+	 * @param elementClass
+	 *            the element class of the generic type
+	 * @return the list with the java objects
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static <T> List<T> toObjectList(final File jsonFile, final Class<T> elementClass)
+		throws IOException
+	{
+		Objects.requireNonNull(jsonFile);
+		Objects.requireNonNull(elementClass);
+		return (List<T>)toObjectCollection(jsonFile, List.class, elementClass);
 	}
 
 }