+++ src/test/java/io/github/astrapi69/json/JsonFileToObjectExtensionsTest.java
@@ -28,8 +28,15 @@ import static org.testng.AssertJUnit.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import io.github.astrapi69.collections.map.MapFactory;
+import io.github.astrapi69.collections.set.SetFactory;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
@@ -53,6 +60,9 @@ public class JsonFileToObjectExtensionsTest
 
 	File jsonDir;
 	File jsonFile;
+	File jsonCollectionFile;
+
+	File jsonMapFile;
 	File jsonListFile;
 
 	@BeforeMethod
@@ -61,6 +71,123 @@ public class JsonFileToObjectExtensionsTest
 		jsonDir = new File(PathFinder.getSrcTestResourcesDir(), "json");
 		jsonFile = new File(jsonDir, "person.json");
 		jsonListFile = new File(jsonDir, "employees.json");
+		jsonMapFile = new File(jsonDir, "map.json");
+		jsonCollectionFile = new File(jsonDir, "collection.json");
+	}
+
+	/**
+	 * Test method for {@link JsonFileToObjectExtensions#toMapObject(File, TypeReference)}
+	 *
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	@Test
+	public void testToMapObjectStringTypeReference() throws IOException
+	{
+		Map<Integer, Integer> actual;
+		Map<Integer, Integer> expected;
+		TypeReference<Map<Integer, Integer>> typeReference;
+
+		// new scenario: try to convert json to integer map
+		typeReference = new TypeReference<Map<Integer, Integer>>()
+		{
+		};
+		actual = JsonFileToObjectExtensions.toMapObject(jsonMapFile, typeReference);
+		expected = MapFactory.newCounterMap(ListFactory.newRangeList(1, 5));
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link JsonFileToObjectExtensions#toMapObject(File, TypeReference, ObjectMapper)}
+	 *
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	@Test
+	public void testToMapObjectStringTypeReferenceObjectMapper() throws IOException
+	{
+		Map<Integer, Integer> actual;
+		Map<Integer, Integer> expected;
+		TypeReference<Map<Integer, Integer>> typeReference;
+
+		// new scenario: try to convert json to integer map
+		typeReference = new TypeReference<Map<Integer, Integer>>()
+		{
+		};
+		actual = JsonFileToObjectExtensions.toMapObject(jsonMapFile, typeReference,
+			ObjectMapperFactory.newObjectMapper());
+		expected = MapFactory.newCounterMap(ListFactory.newRangeList(1, 5));
+		assertEquals(expected, actual);
+	}
+
+
+	/**
+	 * Test method for {@link JsonFileToObjectExtensions#toObjectCollection(File, Class, Class)}
+	 *
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred
+	 */
+	@Test
+	public void testToObjectCollection() throws IOException
+	{
+		boolean actual;
+		boolean expected;
+		Set<Employee> jsonList;
+		Set<Employee> objectList;
+		Employee firstExpected;
+		Employee secondExpected;
+		Employee thirdExpected;
+
+		jsonList = (Set<Employee>)JsonFileToObjectExtensions.toObjectCollection(jsonCollectionFile,
+			LinkedHashSet.class, Employee.class);
+		firstExpected = Employee.builder().person(Person.builder().gender(Gender.FEMALE)
+				.name("Anna").nickname("beast").married(true).about("Ha ha ha...").build()).id("23")
+			.build();
+		secondExpected = Employee.builder().person(Person.builder().gender(Gender.MALE)
+				.name("Andreas").nickname("cute").married(false).about("fine person").build()).id("24")
+			.build();
+		thirdExpected = Employee.builder().person(Person.builder().gender(Gender.FEMALE)
+				.name("Tatjana").nickname("beautiful").married(false).about("Im hot").build()).id("25")
+			.build();
+		objectList = SetFactory.newLinkedHashSet(firstExpected, secondExpected, thirdExpected);
+
+		actual = CollectionExtensions.isEqualCollection(jsonList, objectList);
+		expected = true;
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link JsonFileToObjectExtensions#toObjectList(File, Class)}
+	 *
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	@Test
+	public void testToObjectList() throws IOException
+	{
+		boolean actual;
+		boolean expected;
+		List<Employee> jsonList;
+		List<Employee> objectList;
+		Employee firstExpected;
+		Employee secondExpected;
+		Employee thirdExpected;
+
+		jsonList = JsonFileToObjectExtensions.toObjectList(jsonCollectionFile, Employee.class);
+		firstExpected = Employee.builder().person(Person.builder().gender(Gender.FEMALE)
+				.name("Anna").nickname("beast").married(true).about("Ha ha ha...").build()).id("23")
+			.build();
+		secondExpected = Employee.builder().person(Person.builder().gender(Gender.MALE)
+				.name("Andreas").nickname("cute").married(false).about("fine person").build()).id("24")
+			.build();
+		thirdExpected = Employee.builder().person(Person.builder().gender(Gender.FEMALE)
+				.name("Tatjana").nickname("beautiful").married(false).about("Im hot").build()).id("25")
+			.build();
+		objectList = ListFactory.newArrayList(firstExpected, secondExpected, thirdExpected);
+
+		actual = CollectionExtensions.isEqualCollection(jsonList, objectList);
+		expected = true;
+		assertEquals(expected, actual);
 	}
 
 	/**